requirements.txt
# Core dependencies
aiogram>=3.0.0
aiohttp>=3.8.0
python-dotenv>=1.0.0

# Database
aiosqlite>=0.19.0
sqlalchemy>=2.0.0

# Translation APIs
openai>=1.0.0
langdetect>=1.0.9
googletrans-py>=1.2.3

# Voice processing
pydub>=0.25.1
gtts>=2.3.0
SpeechRecognition>=3.10.0

# PDF generation
reportlab>=4.0.0
pypdf>=3.0.0

# Payment integration
yookassa>=3.0.0

# Utils
pytz>=2023.3
aiofiles>=23.0.0
cryptography>=40.0.0
Pillow>=10.0.0

# Additional dependencies
requests>=2.28.0
asyncio-throttle>=1.0.2